requirements.txt
tensorflow>=2.15.0
keras>=2.15.0
streamlit>=1.30.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
